moved MST into NPC and added visualization
